lua: lsocket_tcp_receive behavior fixed
Receive function behaved inconsistently for the cases where
first we use an integer as a pattern as to tell the function
a specific number of elements we want to read, and then
perform another receive call using '*a' as a pattern.
The latter should have denoted that we want to get all there
is to get, but instead it was possible to get only a part of
data and also in non-sequential order.

Closes #4118